refactor: create graphql folder
